+++ package.json
@@ -31,11 +31,11 @@
     "lint": "eslint src --ext .ts",
     "check": "tsc --noEmit -p tsconfig.json",
     "build:types": "tsc -p tsconfig.types.json",
-    "build:js": "rollup -c",
+    "build:js": "rolldown -c rolldown.config.mjs",
     "build": "run-s build:js build:types",
-    "build:min": "rollup -c",
+    "build:min": "rolldown -c rolldown.config.mjs",
     "build:release": "run-s build build:min copyright",
-    "watch": "rollup -c --watch",
+    "watch": "rolldown -c rolldown.config.mjs --watch",
     "test": "run-s build test:mocha",
     "test:mocha": "mocha \"test/**/*.test.mjs\""
   },
@@ -53,14 +53,7 @@
   },
   "homepage": "https://github.com/cytoscape/cytoscape.js-dagre",
   "devDependencies": {
-    "@babel/core": "^7.29.0",
-    "@babel/preset-env": "^7.29.2",
-    "@babel/preset-typescript": "^7.27.0",
     "@dagrejs/dagre": "^3.0.0",
-    "@rollup/plugin-babel": "^7.0.0",
-    "@rollup/plugin-commonjs": "^29.0.2",
-    "@rollup/plugin-node-resolve": "^16.0.3",
-    "@rollup/plugin-terser": "^1.0.0",
     "@typescript-eslint/eslint-plugin": "^5.62.0",
     "@typescript-eslint/parser": "^5.62.0",
     "camelcase": "^6.0.0",
@@ -73,7 +66,7 @@
     "mocha": "8.1.1",
     "npm-run-all": "^4.1.5",
     "rimraf": "^3.0.2",
-    "rollup": "^4.60.4",
+    "rolldown": "^1.1.3",
     "typescript": "^5.4.0",
     "update": "^0.7.4",
     "updater-license": "^1.0.0"
